New upstream release
Urgency set to high, because under particular
circumstances you could cause a DoS by sending a message with a huge
header.

Include a daily expire script with a setting in /etc/default/dovecot
to enable it. The new script is disabled by default as it requires
some configurations by the system administrator. Thanks to Hans Spaans