d/rules: specify systemd unit dir manually
--with-systemdunitdir=auto requires systemd's pkgconfig file, which is
shipped in the systemd package. Since we do not want to add a B-D on
systemd just for this, specify Debian's default by hand to have the
build system generate and install dovecot.service and dovecot.socket.

Note that this is probably the reason why the systemd units first
appeared in the package in 1:2.2.18-2, even though support should have
been there since 1:2.1.17: 1:2.2.18-2 was the first build after Jessie
was released and the buildds were switched to systemd, thus having
systemd's pkgconfig info file in place.