dovecot-core.postinst: manage 10-ssl.conf using ucf
As #850538 pointed out, not managing 10-ssl.conf causes its own set of
issues. Since we now enable SSL by default, it makes sense to manage
10-ssl.conf under ucf.

Note that we do not provide a list of known-good MD5 sums to ucf, for a
reason: users may have enabled ssl outside 10-ssl.conf; since this is a
behavioural change, we want existing users to be aware of that change
and either use our new 10-ssl.conf, or continue using their own configs.

Closes: #850538